+++ R/tests.R
@@ -54,19 +54,29 @@ testSamples <- function(rows = 10, cols = 5, seed = 12345L) {
   return(samples)
 }
 
-testFeatures <- function(rows = 100, cols = 5, seed = 12345L, numericFeatureID = FALSE) {
+testFeatures <- function(rows = 100, cols = 6, seed = 12345L, numericFeatureID = FALSE) {
   set.seed(seed)
-  features <- matrix(sample(letters, size = rows * (cols - 2), replace = TRUE),
-                    nrow = rows, ncol = cols - 2)
-  colnames(features) <- sprintf("featureVar%02d", seq_len(cols - 2))
-  featureVarNumeric <- sample(1:100, size = rows, replace = TRUE)
-  features <- cbind(featureVarNumeric, features)
   if (numericFeatureID) {
     featureID <- sprintf("%04d", seq_len(rows))
   } else {
     featureID <- sprintf("feature_%04d", seq_len(rows))
   }
-  features <- cbind(customID = featureID, features)
+  secondaryFeatureID <- sprintf("feature_2_%04d", rev(seq_len(rows)))
+  featureVarNumeric <- sample(1:100, size = rows, replace = TRUE)
+  features <- cbind(
+    customID = featureID,
+    secondaryID = secondaryFeatureID,
+    featureVarNumeric
+  )
+  # Fill remaining columns with discrete features
+  nDiscreteFeatures <- cols - ncol(features)
+  discreteFeatures <- matrix(
+    sample(letters, size = rows * nDiscreteFeatures, replace = TRUE),
+    nrow = rows,
+    ncol = nDiscreteFeatures
+  )
+  colnames(discreteFeatures) <- sprintf("featureVar%02d", seq_len(nDiscreteFeatures))
+  features <- cbind(features, discreteFeatures)
   features <- as.data.frame(features, stringsAsFactors = FALSE)
   features <- list(default = features)
   return(features)
@@ -296,19 +306,19 @@ testPlots <- function() {
 
   multiModel_scatterplot <- function(x) {
     ggdf <- data.frame(
-      var1 = x[[1]]$results[,"beta"],
-      var2 = x[[2]]$results[,"beta"]
+      var1 = x[[1]]$results[[1]][,"beta"],
+      var2 = x[[2]]$results[[1]][,"beta"]
     )
     graphics::plot(x = ggdf$var1, y = ggdf$var2)
   }
   assign("multiModel_scatterplot", multiModel_scatterplot, envir = parent.frame())
   multiModel_barplot_sf <- function(x) {
     df <- data.frame(
-      name  = names(x),
-      beta = c(x[[1]]$results[,"beta"], x[[2]]$results[,"beta"])
+      name  = names(x)[1:length(x)-1],
+      beta = c(x[[1]]$results[[1]][,"beta"], x[[2]]$results[[1]][,"beta"])
     )
     graphics::barplot(height = df$beta, names = df$name,
-                      xlab=x[[1]]$results$features_id,
+                      xlab=x[[1]]$results[[1]]$features_id,
                       ylim=c(ifelse(min(df$beta) < 0, min(df$beta)*1.5, -min(df$beta)*1.5),
                              max(df$beta)*1.5))
   }
@@ -393,9 +403,13 @@ testMapping <- function(seed = 12345L, nFeatures = 100,
   model_02_feats[missing_02] <- NA
 
   mapping <- list(
-    model_01 = model_01_feats,
-    model_02 = model_02_feats
+    data.frame(
+      model_01 = model_01_feats,
+      model_02 = model_02_feats,
+      stringsAsFactors = FALSE
+    )
   )
+  names(mapping) <- "default"
 
   return(mapping)
 }