+++ inst/tinytest/testCheck.R
@@ -1,7 +1,7 @@
 # Test checkX() methods
 
 # Setup ------------------------------------------------------------------------
-
+# source(paste0(getwd(), "/inst/tinytest/tinytestSettings.R"))
 source("tinytestSettings.R")
 using(ttdo)
 
@@ -360,10 +360,36 @@ expect_error_xl(
   info = "A single missing value would still be unique. Error if it is found"
 )
 
+resultsDefault <- list(
+  default = list(
+    t1 = data.frame(
+      x = c("a", "b", "c"),
+      y = rnorm(3),
+      stringsAsFactors = FALSE
+    )
+  )
+)
+
+expect_error_xl(
+  addResults(study, results = resultsDefault),
+  'The results cannot be shared using the modelID \"default\"'
+)
+
 # checkEnrichments -------------------------------------------------------------
 
 expect_error_xl(
   addEnrichments(study, enrichments = NULL)
+)
+
+enrichmentDefault <- list(
+  default = list(
+    t1 = NULL
+  )
+)
+
+expect_error_xl(
+  addEnrichments(study, enrichments = enrichmentDefault),
+  "The enrichments cannot be shared using the modelID \"default\""
 )
 
 # checkMetaFeatures ------------------------------------------------------------
@@ -453,37 +479,58 @@ expect_error_xl(
   addMapping(study, mapping = NULL)
 )
 
-## add checks based on check.R
-tempMapping <- list(model_01 = c("feature_01", "feature_02"),
-                    model_02 = c("feature_01", NA))
+tempMapping <- list(data.frame(model_01 = c("feature_01", "feature_02"),
+                               model_02 = c("feature_01", "feature_02"),
+                               stringsAsFactors = FALSE))
+names(tempMapping) <- "default"
 expect_silent_xl(
   addMapping(study, mapping = tempMapping)
 )
 
-tempMapping <- list(model_01 = c("feature_01", "feature_02"),
-                    model_02 = c(NA, NA))
+tempMapping <- list(data.frame(model_01 = c("feature_01", "feature_02"),
+                               model_02 = c("feature_01", NA),
+                               stringsAsFactors = FALSE))
+names(tempMapping) <- "default"
+expect_silent_xl(
+  addMapping(study, mapping = tempMapping)
+)
+
+# check mapping with no named list
+tempMapping <- list(data.frame(model_01 = c("feature_01", "feature_02"),
+                               model_02 = c("feature_01", "feature_02"),
+                               stringsAsFactors = FALSE))
 expect_error_xl(
-  addMapping(study, mapping = tempMapping)
+  addMapping(study, mapping = tempMapping),
+  "The elements of list \"mapping\" must be named"
 )
 
-# check mapping with distinct sizes for elements
-tempMapping <- list(model_01 = c("feature_01", "feature_02"),
-                    model_02 = c("feature_01"))
-expect_silent_xl(
-  addMapping(study, mapping = tempMapping)
+# check mapping with one model having only NAs
+tempMapping <- list(data.frame(model_01 = c("feature_01", "feature_02"),
+                               model_02 = c(NA, NA),
+                               stringsAsFactors = FALSE))
+names(tempMapping) <- "default"
+expect_error_xl(
+  addMapping(study, mapping = tempMapping),
+  "mapping object requires at least one feature per model"
 )
 
 # check mapping with one single element
-tempMapping <- list(model_01 = c("feature_01", "feature_02"))
+tempMapping <- list(data.frame(model_01 = c("feature_01", "feature_02"),
+                               stringsAsFactors = FALSE))
+names(tempMapping) <- "default"
 expect_error_xl(
-  addMapping(study, mapping = tempMapping)
+  addMapping(study, mapping = tempMapping),
+  "mapping object requires at least two models and one feature"
 )
 
 # check mapping features that do not match across models
-tempMapping <- list(model_01 = c("feature_01", "feature_02", NA, NA),
-                    model_02 = c(NA, NA, "feature_05", "feature_06"))
+tempMapping <- list(data.frame(model_01 = c("feature_01", "feature_02", NA, NA),
+                               model_02 = c(NA, NA, "feature_05", "feature_06"),
+                               stringsAsFactors = FALSE))
+names(tempMapping) <- "default"
 expect_error_xl(
-  addMapping(study, mapping = tempMapping)
+  addMapping(study, mapping = tempMapping),
+  "does not present any feature mapped to another model"
 )
 
 # checkBarcodes ----------------------------------------------------------------
@@ -531,3 +578,4 @@ expect_error_xl(
 expect_error_xl(
   addMetaFeaturesLinkouts(study, metaFeaturesLinkouts = NULL)
 )
+